refactor: change folder structure
in this commit i've moved speaker section files from src/components/ to
app/components
