+++ src/app/components/speakers/Speaker.tsx
@@ -1,4 +1,4 @@
-import { speakerInfoType } from "../types";
+import { speakerInfoType } from "./types";
 import Image from "next/image";
 
 const Speaker = ({ speakerInfo }: { speakerInfo: speakerInfoType }) => {
