+++ src/app/components/speakers/types.ts
